Level-1-App/ai-classifier-app/backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
openai==1.82.*
pydantic==2.11.*
python-dotenv==1.1.*
python-multipart==0.0.*
PyMuPDF==1.27.*
python-docx==1.2.*
